+++ exmol/__init__.py
@@ -1,4 +1,6 @@
 from .version import __version__
 from . import stoned
 from .exmol import *
+from .data import *
 from .stoned import sanitize_smiles
+from .plot_utils import moldiff, insert_svg