+++ docs/source/changelog.rst
@@ -1,8 +1,15 @@
 Change Log
 ==========
+
+v0.4.0
+-------------------
+* Refactored code into files
+* Added SVG rewrite so mol structures are SVGs
+
+
 v0.3.2 (2021-09-02)
 -------------------
-* Added SMILES sanitization before generating sample space 
+* Added SMILES sanitization before generating sample space
 
 v0.3.1 (2021-08-26)
 -------------------